bkd/requirements.txt
flask 
openai
python-dotenv

AzureOpenAI





flask-cors